[add]感情分析分岐メソッド

+++ db/migrate/20210304093205_add_magnitude_to_lists.rb
@@ -1,5 +1,5 @@
 class AddMagnitudeToLists < ActiveRecord::Migration[5.2]
   def change
-    add_column :lists, :magnitude, :decimal
+    add_column :lists, :magnitude, :float
   end
 end
